[fixed] the issue with /login and /registration pages for authenticated user
